Moved WallpaperSelector to package "dte.wallpaper"
Because it's a core concept.
